07 Oct 2009

+++ launch/system.c
@@ -843,7 +843,8 @@ void Sys_Error(const char *error, ...)
 	Con_ShowConsole( true );
 	if( Sys.developer >= D_ERROR ) Sys_Print( text );	// print error message
 	else Sys_Print( "Internal engine error\n" );	// don't confuse non-developers with technique stuff
-	Sys.Free(); // kill video
+	if( Sys.app_name == HOST_NORMAL )
+		Sys.Free(); // kill video
 
 	Sys_WaitForQuit();
 	Sys_Exit();
@@ -862,7 +863,9 @@ void Sys_Break(const char *error, ...)
 	Sys.app_state = SYS_ERROR;
 	Con_ShowConsole( true );
 	Sys_Print( text );
-	Sys.Free(); // kill video
+
+	if( Sys.app_name == HOST_NORMAL )
+		Sys.Free(); // kill video
 
 	Sys_WaitForQuit();
 	Sys_Exit();
